requirements.txt
dash
flask
pandas
python-dotenv
rdkit
flask-jwt-extended
dash_bootstrap_components